+++ CHANGES.md
@@ -11,6 +11,7 @@ Changes in 11.6
 - CSS WPadmin theme use OpenSans Semibold font instead of OpenSans Bold in font.css
 - Add vertically align list data option (defaults to false) in ListOutput.fnc.php
 - CSS add .has-input class to list having input (input or select fields) in ListOutput.fnc.php
+- ExplodeDate() use strtotime() in Date.php
 
 Changes in 11.5.1
 -----------------